Update: use uint32 for protobuf

+++ bcs-services/bcs-bscp/cmd/data-service/service/template_space.go
@@ -76,7 +76,7 @@ func (s *Service) ListTemplateSpaces(ctx context.Context, req *pbds.ListTemplate
 	}
 
 	resp := &pbds.ListTemplateSpacesResp{
-		Count:   count,
+		Count:   uint32(count),
 		Details: TemplateSpaces,
 	}
 	return resp, nil
